Make biscuit_auth module init function pub
We have python bindings to both the stock biscuit-auth crate, as well
as our own rust crate that depends on biscuit-auth.  In order for
these to interoperate they all need to be linked together into a
single shared library.  (Or at least, that was the safest way I could
find to do it).

AFAICT python requires that the shared library name match the module
name, so it will not call the module init function in this case.  This
PR makes the `biscuit_auth` module init function `pub` so it can be
called from another crate's module init function.